pyproject.toml
[tool.poetry]
name = "coursework-6"
version = "0.1"
description = ""
authors = ["HAN"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
crispy = "^0.7.4"
django-crispy-forms = "^2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
